Make compatible with Rails 5
In Rails 5, `ActiveRecord::RelationAssociation` no longer delegates to
`Array`, so we cannot use methods like `:<<` to temporarily add
preference objects to the internal `saved_preferences` collection.

The simplest resolution is to use `#save` to immediately persist the
new preference, and then call `#reload` once all of the preferences have
been persisted. This allows the expected behaviour and tests to
continue to pass, at the potential expense of a few more database
queries when setting preferences.