Avoid temporary creation of preference with wrong value
A problem with the previous Rails 5 fix is that it would be possible for
a different process to read a preference after it's been initially saved
but before it's value has actually been updated. We can fix this by not
saving the preference before it's been given the right value.

To do this efficiently, we can use the `find_or_initialize_by` method
introduced in Rails 4. We can also take advantage of the attributes API
to ensure that the preference value is always converted into db format,
which opens us up to potentially even simpler refactorings; the
`update_value!` method now is largely redudant.

In order to to use `find_or_initialize_by`, we've had to drop support
for Rails 3.