Test against modern Ruby versions and fix build
Now that newer versions of some development dependencies (like
sqlite3) are available, we need to be a bit more sophisticated about
how we generate the appraisal gemfiles.

We also need to do a bit of work to get the build (and particularly
the migrations) working for more recent versions of Rails.

Sprockets 4+ requires Ruby 2.5+, but since we're not really interested
in Rails' asset pipeline, we can just fix this at an earlier version.

Finally, we drop Ruby 2.2 from testing, because the `i18n` gem
requires version 2.3 or above, and configuring a test matrix to work
against this is more trouble than it's worth at this point.