Don't attempt to call `=~` on Integer values
In Ruby 3.2, this method is no longer available on integers. We can
avoid the comparison entirely by turning the value for comparison into a
String.

We could also have moved the comparison to `0` above the other lines,
but this would not have handled the situation where we are given another
Integer value (or another class).

Thanks to https://github.com/billabel for this fix!