Add backwards compatibility option for non-binary options
One can set binary true/false preferences the following way:
- set(preference: 0) -> false
- set(preference: 1) -> true
- set(preference: false) -> false
- set(preference: true) -> true
- set(preference: 'false') -> false
- set(preference: 'true') -> true

With a non-binary preference containing `false` and `true` this isn't
possible anymore, as the preference logic will refuse to cast the given
value to a boolean.

This is intended and useful for the most part, but there are cases where
we want to keep the casting behavior for non-binary preferences as well.

Add a `acts_as_binary` option which when set to true will attempt to
cast the given value to a boolean. If the cast succeeds, the boolean is
returned. If not, the given value is returned.

This would allow non-binary preferences to be set like binary
preferences for the `true` and `false` boolean values.